+++ class/GaiaAlpha/App.php
@@ -52,6 +52,7 @@ class App
         $this->router->add('PATCH', '/api/todos/(\d+)', [$todo, 'update']);
         $this->router->add('DELETE', '/api/todos/(\d+)', [$todo, 'delete']);
         $this->router->add('GET', '/api/todos/(\d+)/children', [$todo, 'getChildren']);
+        $this->router->add('POST', '/api/todos/reorder', [$todo, 'reorder']);
 
         // Admin
         $this->router->add('GET', '/api/admin/users', [$admin, 'index']);